+++ apps/server/src/modules/course-synchronization-history/course-synchronization-history.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CourseSynchronizationHistoryMirkoOrmRepo } from './repo/mikro-orm';
 import { COURSE_SYNCHRONIZATION_HISTORY_REPO } from './repo';
-import { CourseSynchronizationHistoryService } from './service';
+import { CourseSynchronizationHistoryService } from './domain';
 
 @Module({
 	providers: [
