+++ apps/server/src/modules/course-synchronization-history/domain/do/course-synchronization-history.ts
