+++ apps/server/src/modules/course-synchronization-history/domain/do/index.ts
