+++ apps/server/src/modules/course-synchronization-history/domain/index.ts
@@ -1,0 +1,2 @@
+export * from './do';
+export * from './service';
