+++ apps/server/src/modules/course-synchronization-history/domain/service/course-synchronization-history.service.spec.ts
@@ -1,7 +1,7 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { Test, TestingModule } from '@nestjs/testing';
-import { COURSE_SYNCHRONIZATION_HISTORY_REPO, CourseSynchronizationHistoryRepo } from '../repo';
-import { courseSynchronizationHistoryFactory } from '../testing';
+import { COURSE_SYNCHRONIZATION_HISTORY_REPO, CourseSynchronizationHistoryRepo } from '../../repo';
+import { courseSynchronizationHistoryFactory } from '../../testing';
 import { CourseSynchronizationHistoryService } from './course-synchronization-history.service';
 
 describe(CourseSynchronizationHistoryService.name, () => {
