+++ apps/server/src/modules/course-synchronization-history/domain/service/index.ts
