+++ apps/server/src/modules/course-synchronization-history/repo/mikro-orm/course-synchronization-history.repo.ts
@@ -1,6 +1,6 @@
 import { EntityData, EntityName } from '@mikro-orm/core';
 import { BaseDomainObjectRepo } from '@shared/repo/base-domain-object.repo';
-import { CourseSynchronizationHistory } from '../../do';
+import { CourseSynchronizationHistory } from '../../domain';
 import { CourseSynchronizationHistoryRepo } from '../course-synchronization-history.repo.interface';
 import { CourseSynchronizationHistoryEntity } from '../entity';
 import { CourseSynchronizationHistoryMapper } from '../mapper';
